Verify HA setup
