+++ cmd/action/verify/cluster/hasetup/command.go
@@ -1,0 +1,45 @@
+package hasetup
+
+import (
+	"github.com/giantswarm/microerror"
+	"github.com/giantswarm/micrologger"
+	"github.com/spf13/cobra"
+)
+
+const (
+	name  = "hasetup"
+	short = "Check if the Tenant Cluster got successfully upgraded to HA."
+	long  = `Check if the Tenant Cluster got successfully upgraded to HA.
+
+	* Check if the all 3 master ready are in ready state.
+	* Return an error if we see other than 3 master nodes.
+`
+)
+
+type Config struct {
+	Logger micrologger.Logger
+}
+
+func New(config Config) (*cobra.Command, error) {
+	if config.Logger == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
+	}
+
+	f := &flag{}
+
+	r := &runner{
+		flag:   f,
+		logger: config.Logger,
+	}
+
+	c := &cobra.Command{
+		Use:   name,
+		Short: short,
+		Long:  long,
+		RunE:  r.Run,
+	}
+
+	f.Init(c)
+
+	return c, nil
+}
