+++ cmd/cl006/zz_generated.command.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/giantswarm/awscnfm/v12/cmd/cl006/ac000"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl006/ac003"
-	"github.com/giantswarm/awscnfm/v12/cmd/cl006/ac004"
 )
 
 const (
@@ -65,20 +64,6 @@ func New(config Config) (*cobra.Command, error) {
 		}
 	}
 
-	var ac004Cmd *cobra.Command
-	{
-		c := ac004.Config{
-			Logger: config.Logger,
-			Stderr: config.Stderr,
-			Stdout: config.Stdout,
-		}
-
-		ac004Cmd, err = ac004.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
 	f := &flag{}
 
 	r := &runner{
@@ -99,7 +84,6 @@ func New(config Config) (*cobra.Command, error) {
 
 	c.AddCommand(ac000Cmd)
 	c.AddCommand(ac003Cmd)
-	c.AddCommand(ac004Cmd)
 
 	return c, nil
 }
